sensor/requirements.txt
paho-mqtt
pyusb
pyserial
boto3
